+++ packages/wallet/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @lucid-evolution/wallet
 
+## 0.1.71
+
+### Patch Changes
+
+- Updated dependencies [[`9503425`](https://github.com/Anastasia-Labs/lucid-evolution/commit/95034257fb3edda5605a76054fdfc70d51640441)]:
+  - @lucid-evolution/utils@0.1.65
+
+## 0.1.70
+
+### Patch Changes
+
+- Updated dependencies [[`2335ef1`](https://github.com/Anastasia-Labs/lucid-evolution/commit/2335ef101f8a5d2a92324c7fd8536d5d8927531e)]:
+  - @lucid-evolution/core-types@0.1.22
+  - @lucid-evolution/sign_data@0.1.25
+  - @lucid-evolution/utils@0.1.64
+
+## 0.1.69
+
+### Patch Changes
+
+- Updated dependencies [[`0342519`](https://github.com/Anastasia-Labs/lucid-evolution/commit/0342519266ca6ff9e96c1e646f7150b78ba095a8)]:
+  - @lucid-evolution/utils@0.1.63
+
 ## 0.1.68
 
 ### Patch Changes